refactor(card_box_panel): split CardBoxPanel into mixin package
Move widgets/panels/card_box_panel.py into a package: frame.py keeps the
class-level template cache and UI construction; properties.py holds the
pure-data color resolver, image-name candidates builder, and text-wrap
helper; handlers.py owns the refresh/assign/update setters, paint/click/
hover callbacks, async image loader, and bitmap renderers.
