+++ widgets/panels/card_box_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Card box panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.card_box_panel.frame import CardBoxPanel
+
+__all__ = ["CardBoxPanel"]
